Update node from version 8 to 10, create new app
Node 10 and the base `create-react-app` configuration
has come a long way since this project was started.

The quickest way to take advantage of the new React and JS features
introduced since version 8 is to:

* Install Node 10
* Create a new, Node 10 React app "harness"
* Migrate the existing application code into the new application harness

Luckily, it's fairly easy to transplant the app's code
from an old project to a new, state of the art app.
This commit handles the first two steps in that process.

This fixes a number of hurdles that I encountered
when trying to install the `styled-components` library,
and generally results in fewer headaches.