Simplify `docker-compose`
Some of the services
weren't playing a useful role,
and could be removed.

I left a small development note in the `docker-compose file,
around possibilities for specifying where data lives.
I think the `DATABASE_URL` envionment variable
is a great de facto standard that the web community has settled on,
and we should work to keep that app <-> data interface
as straightforward and minimal as possible.