Rework application layout to remove scrolling bugs
A few edge cases and inconsistent layout practices
were causing scrollbars to appear where they shouldn't.

We resolved the issues by removing some unneeded layouts and `<div/>`s
from the app, and using grid layouts more strictly.

We added the `<space/>` primitive,
which is handy for taking up a single space in a grid.
Use it with `1fr` in a `grid-template-...`
if you want to take up remaining space.