Overhaul data-fetching logic
The most significant change in this commit
is for fetching data from the server.

Other changes include:

* Move `Assemble.js` -> `Assembly.js`.
* Re-structure the login interface.
* Add a slight border to future calendar dates.
* Remove test suite for `Notes` component.
* Add a timer to the photo upload screen.
* Simplify `src/index.js`
* Translate page titles

There are a number of placeholder changes,
which are marked in the codebase with `TODO` comments.

- - -

## Data Fetching/Saving

Previously, we would need to write a strange mix
of JS and Ruby code to pull the correct records
from the server and process them on the client.

It turned out that most of the code that we were writing
followed the same pattern over and over,
for each of the different data records that we're interested in.

This PR pulls out the `Records` class.
For now, this class is still stored in `src/Assembly.js`,
so it has access to the `network` object defined in that file.

Records offer a simple way to build up queries.
For now, we still translate our queries into ActiveRecord syntax!
We can expand to other query languages as appropriate.

We declare the records that we store in our database,
as:

```javascript
  @observable medication_reports = new Records("MedicationReport", "medication_reports")
  @observable symptom_reports    = new Records("SymptomReport"   , "symptom_reports")
  @observable strip_reports      = new Records("StripReport"     , "strip_reports")
  @observable notes              = new Records("Note"           , "notes")
```

Then, we can take advantage of helper functions to build out our queries:

## Create a record in the database

Original:

```javascript
    this.network.run("cirg")`
      user.medication_reports.create!(
        timestamp: ${JSON.stringify(this.survey_datetime)}
      )
    `
```

Amended:

```javascript
    this.medication_reports.create({ timestamp: this.survey_datetime })
```

## Fetch a list of records from the database

Original:

```javascript
   this.network.watch("cirg")`
      user.notes
    `(response => {
      response
        .json()
        .then(action(notes => {
          this.notes = notes
        }))
        .catch(() => {})
```

Amended:

```javascript
    this.notes.watch()
    this.notes.records // => [{...}, {...}]
```